changes for reloading routes after user login
