+++ frontend/src/App.tsx
@@ -10,29 +10,6 @@ import SideBar from './app/components/navigation/SideBar';
 import '@bcgov/bc-sans/css/BCSans.css';
 
 function App() {
-  const { isAuthenticated, signinSilent, events, user, signoutSilent } =
-    useAuth();
-
-  const tryTokenRefresh = useCallback(() => {
-    signinSilent().then((data) => {
-      // Refresh failed, this usually means that refresh token is invalid or expired.
-      // Sign out the user and clear the token data in this case.
-      if (data === null) {
-        signoutSilent();
-      }
-    });
-  }, [signinSilent, signoutSilent]);
-
-  useEffect(() => {
-    if (user?.expired) {
-      tryTokenRefresh();
-    }
-    // the `return` is important - addAccessTokenExpiring() returns a cleanup function
-    return events.addAccessTokenExpiring(() => {
-      tryTokenRefresh();
-    });
-  }, [events, isAuthenticated, user, tryTokenRefresh]);
-
   return (
     <div className="container-fluid p-0">
       <Header />
