+++ frontend/src/index.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useCallback, useEffect } from 'react';
 import ReactDOM from 'react-dom/client';
 import {
   ApolloClient,
@@ -13,12 +13,12 @@ import reportWebVitals from './reportWebVitals';
 import { store } from './app/Store';
 import { Provider } from 'react-redux';
 import 'bootstrap/dist/css/bootstrap.min.css';
-import { AuthProvider } from 'react-oidc-context';
+import { AuthProvider, useAuth } from 'react-oidc-context';
 import { UserManagerSettings } from 'oidc-client-ts';
 import { getClientSettings } from './app/auth/UserManagerSetting';
 import { RouterProvider } from 'react-router-dom';
 import siteRouter from './app/routes/Routes';
-import { getUser } from './app/helpers/utility';
+import { getLoggedInUserType, getUser } from './app/helpers/utility';
 import { API, GRAPHQL } from './app/helpers/endpoints';
 
 const root = ReactDOM.createRoot(
@@ -46,14 +46,49 @@ const client = new ApolloClient({
 });
 
 const authOptions: UserManagerSettings = getClientSettings();
+
+function AppWrapper() {
+  const [userType, setUserType] = React.useState(getLoggedInUserType());
+  const { isAuthenticated, signinSilent, events, user, signoutSilent } =
+    useAuth();
+
+  useEffect(() => {
+    setUserType(getLoggedInUserType());
+  }, [user]);
+
+  const tryTokenRefresh = useCallback(() => {
+    signinSilent().then((data) => {
+      // Refresh failed, this usually means that refresh token is invalid or expired.
+      // Sign out the user and clear the token data in this case.
+      if (data === null) {
+        signoutSilent();
+      }
+    });
+  }, [signinSilent, signoutSilent]);
+
+  useEffect(() => {
+    if (user?.expired) {
+      tryTokenRefresh();
+    }
+    // the `return` is important - addAccessTokenExpiring() returns a cleanup function
+    return events.addAccessTokenExpiring(() => {
+      tryTokenRefresh();
+    });
+  }, [events, isAuthenticated, user, tryTokenRefresh]);
+
+  return (
+    <ApolloProvider client={client}>
+      <Provider store={store}>
+        <RouterProvider router={siteRouter(userType)} />
+      </Provider>
+    </ApolloProvider>
+  );
+}
+
 root.render(
   <React.StrictMode>
     <AuthProvider {...authOptions}>
-      <ApolloProvider client={client}>
-        <Provider store={store}>
-          <RouterProvider router={siteRouter} />
-        </Provider>
-      </ApolloProvider>
+      <AppWrapper />
     </AuthProvider>
   </React.StrictMode>,
 );
